pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepflow"
version = "0.1.0"
description = "DeepFlow runtime, ACP bridge, and Langflow integration for Deep Agents."
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "deepagents==0.4.11",
    "deepagents-acp>=0.0.4,<1.0.0",
    "ddgs>=9.0.0,<10.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "langchain-anthropic>=1.3.4,<2.0.0",
    "langchain-deepseek>=1.0.0,<2.0.0",
    "langchain-google-genai>=4.2.0,<5.0.0",
    "langchain-ollama>=1.0.0,<2.0.0",
    "langchain-openai>=1.1.8,<2.0.0",
    "langchain-openrouter>=0.1.0,<1.0.0",
    "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
    "markdownify>=1.2.0,<2.0.0",
    "mcp[cli]>=1.20.0,<2.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "tavily-python>=0.7.0,<1.0.0",
    "typer>=0.16.0,<1.0.0",
    "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0,<9.0.0",
]

[project.scripts]
deepflow = "deepflow_runtime.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/deepflow_runtime"]
